+++ src/casia/isiteam/recommendsystem/model/Recommendation.java
@@ -6,11 +6,11 @@ package casia.isiteam.recommendsystem.model;
 public class Recommendation {
 
     // 推荐项id
-    private int id;
+    private long id;
     // 用户id
-    private int user_id;
+    private long user_id;
     // 新闻id
-    private int news_id;
+    private long news_id;
     // 生成时间
     private String derive_time;
     // 用户反馈 0 未浏览 1 已浏览
@@ -21,7 +21,7 @@ public class Recommendation {
     public Recommendation() {
     }
 
-    public Recommendation(int id, int user_id, int news_id, String derive_time, int feedback, int derive_algorithm) {
+    public Recommendation(long id, long user_id, long news_id, String derive_time, int feedback, int derive_algorithm) {
         this.id = id;
         this.user_id = user_id;
         this.news_id = news_id;
@@ -30,27 +30,27 @@ public class Recommendation {
         this.derive_algorithm = derive_algorithm;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public int getUser_id() {
+    public long getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(int user_id) {
+    public void setUser_id(long user_id) {
         this.user_id = user_id;
     }
 
-    public int getNews_id() {
+    public long getNews_id() {
         return news_id;
     }
 
-    public void setNews_id(int news_id) {
+    public void setNews_id(long news_id) {
         this.news_id = news_id;
     }
